requirements.txt
batchgenerators==0.25
connected_components_3d==3.17.0
cudf==0.6.1.post1
cugraph==0.6.1.post1
cuml==0.6.1.post1
dicom_utils==0.1.0
fastapi==0.115.11
fastcore==1.6.3
grad_cam==1.5.3
gudhi==3.10.1
h5py==3.12.1
ipdb==0.13.13
itk==5.4.2.post1
lightning==2.5.0.post0
lightning_utilities==0.11.9
matplotlib==3.10.1
monai==1.4.0
neptune==1.13.0
numpy==2.2.3
nurbspy==1.1.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
paramiko==3.5.0
Pillow==11.1.0
plotly==6.0.0
prefetch_generator==1.0.3
psutil==6.0.0
pydicom==2.3.0
pytest==8.3.3
PyYAML==6.0.2
ray==2.36.1
reportlab==4.2.5
Requests==2.32.3
roboflow==1.1.49
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.1.0
SimpleITK==2.4.1
skimage==0.0
supervision==0.25.1
torch==2.5.1+cu121
torchinfo==1.8.0
torchio==0.20.4
torchvision==0.20.1+cu121
tqdm==4.66.5
ultralytics==8.3.44
utilz.egg==info
xnat==0.7.0
